backend/requirements.txt
# Core API framework
fastapi==0.115.0

# ASGI server
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
asyncpg==0.29.0

# Settings management
pydantic-settings==2.5.2

# WebSocket client (for collectors)
websockets==13.1

# Claude API client (for analysis worker)
anthropic>=0.40.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
